requirements.txt
streamlit
pandas
sqlalchemy
pyodbc
python-dotenv
altair